kinetics/preprocess/requirements.txt
moviepy==1.0.3
numpy==2.0.1
Pillow==10.4.0
tqdm==4.66.5
# opencv-python==4.10.0.84
# opencv-contrib-python==4.10.0.84